requirements.txt
aiosignal==1.3.1
annotated-types==0.6.0
arrow==1.3.0
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
black==23.10.1
blinker==1.6.3
blis==0.7.11
cachetools==5.3.1
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cloudpathlib==0.15.1
confection==0.1.3
contourpy==1.1.1
cycler==0.12.1
cymem==2.0.8
decorator==5.1.1
dnspython==2.4.2
filelock==3.12.4
Flask==3.0.0
Flask-Cors==4.0.0
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==3.1.1
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.9.2
google-api-core==2.12.0
google-auth==2.23.3
google-cloud-core==2.3.3
google-cloud-speech==2.21.0
google-cloud-translate==3.12.0
googleapis-common-protos==1.61.0
grpcio==1.59.0
grpcio-status==1.59.0
huggingface-hub==0.17.3
ics==0.7.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
langcodes==3.3.0
langdetect==1.0.9
lazy_loader==0.3
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.8.0
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
mysqlclient==2.2.0
networkx==3.1
nltk==3.8.1
numpy==1.26.1
packaging==23.2
pathspec==0.11.2
pathy==0.10.2
Pillow==10.0.1
platformdirs==3.11.0
pooch==1.7.0
preshed==3.0.9
proto-plus==1.22.3
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
pygame==2.5.2
pymongo==4.5.0
PyMySQL==1.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
python-docx==1.0.1
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
rsa==4.9
safetensors==0.4.0
scikit-learn==1.3.1
scipy==1.11.3
setuptools==68.2.2
simpleaudio==1.0.4
six==1.16.0
smart-open==6.4.0
sounddevice==0.4.6
soundfile==0.12.1
soxr==0.3.7
spacy==3.7.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.0
SQLAlchemy==2.0.22
srsly==2.4.8
sympy==1.12
TatSu==5.8.3
thinc==8.2.1
threadpoolctl==3.2.0
tk==0.1.0
tokenizers==0.14.1
tqdm==4.66.1
transformers==4.34.0
typer==0.9.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
urllib3==2.0.6
wasabi==1.1.2
weasel==0.3.2
Werkzeug==3.0.0
whisper==1.1.10
yarl==1.9.2
